Restructure directories and imports. (#6)
* Restructure directories and imports.

* Remove redundant path/filename element in views.
