+++ src/app/models/user.ts
@@ -1,6 +1,6 @@
 // A client user, simulating login via Google or OAuth in a browser.
 
-import {PermissionLevel} from './permission-level';
+import {PermissionLevel} from 'app/models/permission-level';
 
 export class User {
   id: number;
