+++ src/app/views/select-repository/component.ts
@@ -7,12 +7,12 @@
 import {Component, OnInit} from '@angular/core';
 import {Router} from '@angular/router';
 
-import {Repository} from './repository';
-import {RepositoryService} from './repository.service';
-import {User} from './user';
-import {UserService} from './user.service';
+import {Repository} from 'app/models/repository';
+import {RepositoryService} from 'app/services/repository.service';
+import {User} from 'app/models/user';
+import {UserService} from 'app/services/user.service';
 
-@Component({templateUrl: './select-repository.component.html'})
+@Component({templateUrl: './component.html'})
 export class SelectRepositoryComponent implements OnInit {
   repositories: Repository[] = [];
   user: User;
