+++ src/main.ts
@@ -2,6 +2,6 @@
 
 import {platformBrowserDynamic} from '@angular/platform-browser-dynamic';
 
-import {AppModule} from './app/app.module';
+import {AppModule} from 'app/app.module';
 
 platformBrowserDynamic().bootstrapModule(AppModule);
